moving util contents to utilities and renaming sharedMiddlewares to middlewares
