+++ src/middlewares/protectRoutes.middleware.ts
