+++ src/utilities/catchAsync.ts
@@ -1,4 +1,4 @@
-import { RequestHandler, Request, Response, NextFunction } from "express";
+import { Request, Response, NextFunction } from "express";
 
 const catchAsync = (
   fn: (req: Request, res: Response, next: NextFunction) => Promise<void>
